List all configured DNS servers, not just 2 (#1866)
This can make the widget much taller for the rare cases where someone
has a lot of servers configured. I think the extra height it worth
correctness.

In the future, I'd like to consider a setting to limit the number or
intelligent widget sizing where a widget can consider the available room,
and adjust its output.

This PR is an alternative solution to #1865.